[QA] BOAC-2103 BOAC-2091 Show users their supplied attachment filenames ... but avoid S3 name collisions internally.
